fix.solucionar routing dashboard
