+++ frontend-vinculacion/src/app/app-routing.module.ts
@@ -2,16 +2,12 @@ import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 import { NopagesfoundComponent } from './nopagesfound/nopagesfound.component';
 
-
-
 const routes: Routes= [
-  { path: '', redirectTo: '/pages/solicitud', pathMatch: 'full' },
+  { path: '', redirectTo: '/pages/dashboard', pathMatch: 'full' },
   {path: 'pages', loadChildren:() => import('./pages/pages.module').then(m => m.PagesModule)},
   {path: 'shared', loadChildren:() => import('./shared/shared.module').then(m => m.SharedModule)},
   { path: '**', component: NopagesfoundComponent },
 ]
-
-
 
 @NgModule({
   imports: [RouterModule.forRoot(routes)],
